requirements.txt
pandas
seaborn
numpy
matplotlip
scikit-learn
xgboost
dill
django==5.1.3
djangorestframework==3.15.2
python-dotenv==1.0.0
python-whois
googlesearch-python
beautifulsoup4
tldextract
django-cors-headers